UI: Added auth login and signup page
