+++ front-end/components/Navbar/index.js
@@ -1,0 +1,9 @@
+const Navbar = () => {
+  return (
+    <nav>
+      <h1>Ingenious</h1>
+    </nav>
+  );
+};
+
+export default Navbar;
