+++ front-end/pages/codeground/index.js
@@ -1,0 +1,65 @@
+import Head from "next/head";
+import Editor from "@monaco-editor/react";
+import { useEffect, useRef, useState } from "react";
+import axios from "axios";
+
+export default function Home() {
+  const editorRef = useRef(null);
+
+  const [language, setLanguage] = useState("python");
+
+  const handleEditorDidMount = (editor, monaco) => {
+    editorRef.current = editor;
+  };
+
+  const showValue = () => {
+    console.log(editorRef.current.getValue());
+  };
+
+  const handleSubmit = async () => {
+    try {
+      const response = await axios.post("http://localhost:5000/compile", {
+        language,
+        code: editorRef.current.getValue(),
+      });
+
+      console.log(response.data);
+    } catch (err) {
+      console.log(err);
+    }
+  };
+
+  useEffect(() => {
+    console.log(language);
+  }, [language]);
+
+  return (
+    <div>
+      <Head>
+        <title>Compiler Project X</title>
+        <meta name="description" content="Generated by create next app" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+
+      <button onClick={showValue}>Show Value</button>
+      <Editor
+        height="90vh"
+        width={"80vw"}
+        theme="vs-dark"
+        onMount={handleEditorDidMount}
+        defaultLanguage="cpp"
+        defaultValue="#Some comment"
+      />
+      <select
+        onChange={(event) => {
+          setLanguage(event.target.value);
+        }}
+      >
+        <option value={"python"}>Python</option>
+        <option value={"cpp"}>C++</option>
+        <option value={"javascript"}>Javascript</option>
+      </select>
+      <button onClick={handleSubmit}>Run Code</button>
+    </div>
+  );
+}
